requirements.txt
streamlit
spacy
langdetect
pdfplumber
python-docx
nltk
